Modernize for Dalamud API 15 (.NET 10)
Revive the plugin against the current Dalamud SDK (11 -> 15, net8 -> net10):

- csproj: bump SDK + target framework; drop the stale hand-rolled <Reference>
  block (including the removed ImGui.NET/ImGuiScene); pin SDK-15 output paths.
- Migrate ImGuiNET -> Dalamud.Bindings.ImGui across the UI.
- API drift: IClientState.LocalPlayer -> IObjectTable.LocalPlayer;
  TerritoryType/TerritoryChanged ushort -> uint; adopt the new single-object
  IHandleableChatMessage chat delegate.
- Remove defunct backends: LootHTTP, LootDiscord, and the DiscordBridge IPC
  consumer (the DiscordBridge plugin no longer exists), plus their config and
  settings tabs.
- Manifest/distribution: update Kapture.yaml (API 15, icon, changelog); add a
  self-serve repo.json; document install paths in the README.
- Tests: retarget to net10, bump xunit/test-sdk, stage Dalamud libs for the
  test host, and fix a mock stub that wiped PlayerName. 93 passing.
- Keep a loadable DLL in the staged folder by re-extracting latest.zip in the
  project's DalamudPackager.targets override.

+++ src/Kapture/Kapture/Extensions/ContentExtension.cs
@@ -18,7 +18,7 @@ namespace Kapture.Kapture.Extensions
         /// <param name="value">data manager.</param>
         /// <param name="territoryType">territory type id.</param>
         /// <returns>content id or zero if not in content.</returns>
-        public static uint ContentId(this IDataManager value, ushort territoryType) => GetContentId(value, territoryType);
+        public static uint ContentId(this IDataManager value, uint territoryType) => GetContentId(value, territoryType);
 
         /// <summary>
         /// Get content name.
@@ -26,7 +26,7 @@ namespace Kapture.Kapture.Extensions
         /// <param name="value">data manager.</param>
         /// <param name="territoryType">territory type id.</param>
         /// <returns>content name.</returns>
-        public static string ContentName(this IDataManager value, ushort territoryType)
+        public static string ContentName(this IDataManager value, uint territoryType)
         {
             var contentId = GetContentId(value, territoryType);
             if (contentId == 0)
@@ -43,7 +43,7 @@ namespace Kapture.Kapture.Extensions
         /// <param name="value">data manager.</param>
         /// <param name="territoryType">territory type id.</param>
         /// <returns>indicator whether local player is in content.</returns>
-        public static bool InContent(this IDataManager value, ushort territoryType) => GetContentId(value, territoryType) != 0;
+        public static bool InContent(this IDataManager value, uint territoryType) => GetContentId(value, territoryType) != 0;
 
         /// <summary>
         /// Gets indicator whether current territory is high-end duty content.
@@ -51,7 +51,7 @@ namespace Kapture.Kapture.Extensions
         /// <param name="value">data manager.</param>
         /// <param name="territoryType">territory type id.</param>
         /// <returns>indicator whether local player is in high-end duty content.</returns>
-        public static bool InHighEndDuty(this IDataManager value, ushort territoryType)
+        public static bool InHighEndDuty(this IDataManager value, uint territoryType)
         {
             var contentId = GetContentId(value, territoryType);
             if (contentId == 0)
@@ -179,7 +179,7 @@ namespace Kapture.Kapture.Extensions
             return world?.DataCenter.RowId == 13;
         }
 
-        private static uint GetContentId(IDataManager value, ushort territoryType)
+        private static uint GetContentId(IDataManager value, uint territoryType)
         {
             var cfcs = value.GetExcelSheet<ContentFinderCondition>();
             var cfcResult = cfcs.TryGetFirst(c => c.TerritoryType.RowId == territoryType, out var cfc);
